pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "protein-affinity-gpu"
dynamic = ["version"]
description = "Protein binding affinity prediction with CPU, JAX, and tinygrad backends"
requires-python = ">=3.10"
authors = [{ name = "mcale6" }]
dependencies = [
    "biopython",
    "prodigy-prot",
    "numpy>=1.23,<3.0"
]
readme = "README.md"
license = { file = "LICENSE" }

[project.optional-dependencies]
compare = ["freesasa", "jax", "jaxlib", "tinygrad", "matplotlib"]
jax = ["jax", "jaxlib"]
tinygrad = ["tinygrad"]
bench = ["matplotlib", "pandas"]
dev = ["build", "pytest>=8.0", "ruff>=0.6"]

[project.scripts]
protein-affinity-predict = "protein_affinity_gpu.cli.predict:main"

[tool.hatch.build.targets.wheel]
packages = ["src/protein_affinity_gpu"]

[tool.hatch.build.targets.sdist]
include = [
    ".github/workflows/ci.yml",
    "README.md",
    "LICENSE",
    "benchmarks/**",
    "pyproject.toml",
    "src/protein_affinity_gpu/**",
    "tests/**",
    "update_pkg.sh",
]

[tool.hatch.version]
path = "src/protein_affinity_gpu/version.py"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py310"
